Completely changed how XnoeOS boots, organised the code somewhat, renamed kernel32 to kernel, now uses the 9 sectors after the boot sector to store a 32 bit stage2 loader. No longer spends unnecessary time in real mode. Updated atapio to properly account for the reserved sector count.

+++ src/kernel/idt.c
@@ -3,7 +3,7 @@
 GateEntry idt[256];
 
 void set_entry(uint8_t interrupt_number, uint16_t code_segment, void* handler, uint8_t type) {
-  uint32_t handler_addr = (uint16_t)handler;
+  uint32_t handler_addr = (uint32_t)handler;
   uint16_t* handler_halves = (uint16_t*)&handler_addr;
   idt[interrupt_number] = (GateEntry){
     .offset_low = handler_halves[0],
